feat: add artists filtering input
