+++ src/app/Scenes/MyCollection/Components/MyCollectionCollectedArtistsView.tsx
@@ -1,14 +1,11 @@
 import { Flex, Spacer, Spinner } from "@artsy/palette-mobile"
 import { MyCollectionCollectedArtistsView_me$key } from "__generated__/MyCollectionCollectedArtistsView_me.graphql"
-import { MyCollectionCollectedArtistGridItem } from "app/Scenes/MyCollection/Components/MyCollectionCollectedArtistGridItem"
+import { MyCollectionArtistFilters } from "app/Scenes/MyCollection/Components/MyCollectionArtistFiltersStickyTab"
+import { MyCollectionArtworksKeywordStore } from "app/Scenes/MyCollection/Components/MyCollectionArtworksKeywordStore"
 import { MyCollectionCollectedArtistItem } from "app/Scenes/MyCollection/Components/MyCollectionCollectedArtistItem"
-import { ViewAsIcons } from "app/Scenes/MyCollection/Components/MyCollectionSearchBar"
-import { ViewOption } from "app/Scenes/Search/UserPrefsModel"
-import { GlobalStore } from "app/store/GlobalStore"
 import { extractNodes } from "app/utils/extractNodes"
-import { isPad } from "app/utils/hardware"
 import { useState } from "react"
-import { FlatList, LayoutAnimation, RefreshControl } from "react-native"
+import { FlatList, RefreshControl } from "react-native"
 import { graphql, usePaginationFragment } from "react-relay"
 
 interface MyCollectionCollectedArtistsViewProps {
@@ -19,12 +16,13 @@ export const MyCollectionCollectedArtistsView: React.FC<MyCollectionCollectedArt
   me,
 }) => {
   const [refreshing, setRefreshing] = useState(false)
-  const viewOption = GlobalStore.useAppState((state) => state.userPrefs.artworkViewOption)
-  const isAPad = isPad()
+
   const { data, hasNext, loadNext, isLoadingNext, refetch } = usePaginationFragment(
     collectedArtistsPaginationFragment,
     me
   )
+
+  const keyword = MyCollectionArtworksKeywordStore.useStoreState((state) => state.keyword)
 
   const handleLoadMore = () => {
     if (!hasNext || isLoadingNext) {
@@ -32,15 +30,6 @@ export const MyCollectionCollectedArtistsView: React.FC<MyCollectionCollectedArt
     }
 
     loadNext(10)
-  }
-
-  const onViewOptionChange = (selectedViewOption: ViewOption) => {
-    LayoutAnimation.configureNext({
-      ...LayoutAnimation.Presets.linear,
-      duration: 100,
-    })
-
-    GlobalStore.actions.userPrefs.setArtistViewOption(selectedViewOption)
   }
 
   const handleRefresh = () => {
@@ -63,6 +52,24 @@ export const MyCollectionCollectedArtistsView: React.FC<MyCollectionCollectedArt
   }
 
   const artistsAndArtworksCount = data.myCollectionInfo?.collectedArtistsConnection?.edges
+    ?.filter((artist) => {
+      return (
+        artist?.node?.name
+          ?.toLowerCase()
+          // Make sure to remove accents and diacritics for better comparisons
+          // Salvador Dalí -> Salvador Dali
+          // Édouard Manet -> Edouard Manet
+          .normalize("NFD")
+          .replace(/[\u0300-\u036f]/g, "")
+          .includes(
+            keyword
+              .toLowerCase()
+              .trim()
+              .normalize("NFD")
+              .replace(/[\u0300-\u036f]/g, "")
+          )
+      )
+    })
     ?.filter((edge) => edge !== null && edge.node !== null)
     .map((edge) => {
       return {
@@ -72,64 +79,37 @@ export const MyCollectionCollectedArtistsView: React.FC<MyCollectionCollectedArt
     })
 
   return (
-    <>
-      <Flex alignItems="flex-end" pb={2} px={2}>
-        <ViewAsIcons onViewOptionChange={onViewOptionChange} viewOption={viewOption} />
-      </Flex>
-      {viewOption === "grid" ? (
-        <FlatList
-          data={artistsAndArtworksCount}
-          renderItem={({ item }) => {
-            return (
-              <MyCollectionCollectedArtistGridItem
-                artworksCount={item.artworksCount}
-                // castiing this type as ts was not able to infer it
-                artist={item.artist!}
-                // castiing this type as ts was not able to infer it
-                key={item.artist!.internalID}
-              />
-            )
-          }}
-          key="grid"
-          keyExtractor={(item) => "grid" + item.artist!.internalID}
-          numColumns={isAPad ? 5 : 2}
-          onEndReached={handleLoadMore}
-          ListFooterComponent={!!hasNext ? <LoadingIndicator /> : <Spacer y={2} />}
-          ItemSeparatorComponent={() => <Spacer y={2} x={4} />}
-          refreshControl={
-            !__TEST__ ? (
-              <RefreshControl onRefresh={handleRefresh} refreshing={refreshing} />
-            ) : undefined
-          }
-        />
-      ) : (
-        <FlatList
-          data={artistsAndArtworksCount}
-          key="list"
-          keyExtractor={(item) => "list" + item.artist!.internalID}
-          renderItem={({ item }) => {
-            return (
-              <MyCollectionCollectedArtistItem
-                artworksCount={item.artworksCount}
-                // casting this type because typescript was not able to infer it correctly
-                artist={item.artist!}
-                // casting this type because typescript was not able to infer it correctly
-                key={item.artist!.internalID}
-                compact
-              />
-            )
-          }}
-          onEndReached={handleLoadMore}
-          ListFooterComponent={!!hasNext ? <LoadingIndicator /> : <Spacer y={2} />}
-          ItemSeparatorComponent={() => <Spacer y={2} />}
-          refreshControl={
-            !__TEST__ ? (
-              <RefreshControl onRefresh={handleRefresh} refreshing={refreshing} />
-            ) : undefined
-          }
-        />
-      )}
-    </>
+    <Flex>
+      <MyCollectionArtistFilters />
+
+      <Spacer y={1} />
+
+      <FlatList
+        data={artistsAndArtworksCount}
+        key="list"
+        keyExtractor={(item) => "list" + item.artist!.internalID}
+        renderItem={({ item }) => {
+          return (
+            <MyCollectionCollectedArtistItem
+              artworksCount={item.artworksCount}
+              // casting this type because typescript was not able to infer it correctly
+              artist={item.artist!}
+              // casting this type because typescript was not able to infer it correctly
+              key={item.artist!.internalID}
+              compact
+            />
+          )
+        }}
+        onEndReached={handleLoadMore}
+        ListFooterComponent={!!hasNext ? <LoadingIndicator /> : <Spacer y={2} />}
+        ItemSeparatorComponent={() => <Spacer y={2} />}
+        refreshControl={
+          !__TEST__ ? (
+            <RefreshControl onRefresh={handleRefresh} refreshing={refreshing} />
+          ) : undefined
+        }
+      />
+    </Flex>
   )
 }
 
@@ -156,6 +136,7 @@ const collectedArtistsPaginationFragment = graphql`
           artworksCount
           node {
             internalID
+            name
             ...MyCollectionCollectedArtistItem_artist
             ...MyCollectionCollectedArtistGridItem_artist
           }
